ai-chat-tag-chat-reconf-v1.2-v2.5/requirements.txt
gunicorn==19.7.1
Django==2.0.2
pytz==2017.3
djangorestframework==3.7.7
pymysql==0.8.0
kazoo==2.4.0
requests==2.18.4
django-redis==4.9.0
DBUtils==1.2
pymongo==3.6.0